A lot of fixes for 6.8 beta (#862)
* Improves boundary building with track extension

Adds an "Auto Find" feature to the boundary builder that extends tracks by 50 meters to ensure intersections, particularly for AB lines, are found just beyond original endpoints.

The track extension functionality adds extra length to both ends of a track polyline to help ensure intersections are found.

The extend/shrink track point buttons now shift by 10 instead of 5.

* Refactors track extension logic

Improves the accuracy of track extension by using GeoCoords and GeoDelta calculations.

This ensures that the track extensions are calculated correctly regardless of coordinate system or projection.
Also, the code is simplified by removing redundant checks.

* Refactors goal point distance calculation

Removes redundant logic that adjusted the look-ahead distance based on the auto-steer setting. This logic was causing inconsistencies in goal point distance calculation.

* Improves FormNudge appearance and behavior

Refactors FormNudge to remove reliance on window size,
and enables dragging from any empty client area within the form.

Prevents maximizing on double click and via system commands.

Adds drop shadow effect for a better visual experience.

* Updates Lithuanian translations (#873)

* Updates Lithuanian translations

Refines Lithuanian translations for improved clarity and accuracy.

This update includes corrections and additions to various terms and phrases
within the application, aiming to provide a more intuitive user experience
for Lithuanian-speaking users.

* Remove unused keys from Lithuanian translation

* Restore previous translation for gsDeleteAllContoursAndSections

---------

Co-authored-by: FantasyTeddy <matthiasbaer@bluewin.ch>

* Delays hotkey filter until UI is ready

Prevents hotkey processing before the UI is fully initialized.
This avoids potential issues with controls not being ready
when a hotkey is pressed early in the application lifecycle.

The hotkey filter is now enabled only after the main form
has been shown and the UI is considered "ready".
Fixes #866

* Initializes hotkey filter correctly

Ensures the hotkey filter is initialized properly
before being added as a message filter. This prevents
potential issues with control initialization order.

* Enhances hotkey handling and initialization

Refactors hotkey processing to improve reliability and prevent errors during initialization.

Guards against null reference exceptions and ensures that hotkeys are only processed when the UI is fully initialized and hotkey mappings are available.

Adds documentation to clarify the purpose and functionality of key methods.

* fix Terms and Conditions

Fixes typos and improves phrasing in the translation strings for better clarity and consistency.
Fixes #867

* Restore original code for mousewheel zoom

* Get original code running again.

* Move zoom code to Camera class

* Correct size of RollIndicator. (Was half size since Feb 17)

* Fix exception when opening Bing map (#880)

The System.Data.Interop.dll was not copied properly to the output folder. With version 2.0.1 of System.Data.SQLite, that is not an issue anymore, because it does not rely on that DLL anymore.

* retuned the new sounds based on beta testers comments (#881)

* Adds overwrite confirmation for profile creation (#884)

* Adds overwrite confirmation for profile creation

Prompts the user for confirmation before overwriting an existing profile.
Adds an extra confirmation for resetting to an empty profile, warning the user about potential data loss.
Handles empty profile names gracefully, preventing errors.

* Close confirmation, Advanced Settings Panel, added StartMinimized option  (#883)

* Adds start minimized option

Adds a "Start Minimized" option to the AgIO form.

This allows the user to configure the application to start in a minimized state, saving space on the taskbar.
The option is persisted to the application settings.

* Adds confirmation dialog on exit

Adds a confirmation dialog when the user attempts to exit the application.

This prevents accidental closure of the application by requiring the user to explicitly confirm their intention to exit.
Also moves the start minimized checkbox and the minimize button to the top right of the window.

* Adds advanced settings form

Adds a new form for advanced settings, including options for
starting minimized and auto-running GPS-Out.
Removes the checkboxes controlling these settings from the main form and
places them in the new advanced settings form.
This commit also adds a minimize icon.

* UI Fixes + AgShareDownloader fix (#889)

* UI Fixes

Headland distance was at wrong place.
added some logic if headland distance isnt used to move Trackinfo up.

* Set timeout on HttpClient directly

---------

Co-authored-by: FantasyTeddy <matthiasbaer@bluewin.ch>

* Fix ABDraw/QuicAB (#890)

* Fix ABDraw/QuicAB

Users can't continue saving an AB Line if A+B hasnt been pressed first (cancel works)

* Headland indicator improvement

Also check if Headland is turned off by user to move Trackinfo down.
Better visibility for Headland Indicator if close to headland (box going red)

---------

Co-authored-by: FantasyTeddy <matthiasbaer@bluewin.ch>
Co-authored-by: Joost van Rooijen <joost.rzz@gmail.com>
Co-authored-by: J005t67 <156546745+J005t67@users.noreply.github.com>